backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
PyMySQL==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
email-validator==2.1.0
reportlab==4.0.7
Pillow==10.1.0
Werkzeug==3.0.1
gunicorn==21.2.0
